Fixed integer passing problem for species similarity threshold

+++ src/mtbin.cpp
@@ -110,7 +110,7 @@ unsigned int MtBin::alive() const {
   return startofdead;
 }
 
-void MtBin::breed(double t, int species_sim_thresh){
+void MtBin::breed(double t, double species_sim_thresh){
   if(startofdead==0) return;       //No one is alive here. There can be no breeding.
 
   unsigned int maxalive=kkap(t);   //Current carrying capacity of the bin
